+++ changelog.md
@@ -9,3 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2020-03-29
 ### Added
 - All new files uploaded to GitHub for version 1.0.0
+
+## [1.1.0] - 2022-06-20
+### Added
+- Change to Bootstrap 5
+- Update setup script to add default manufacturers 
+- Update dashboard layout
+- Update scripts to remove /root category from part insert and edit pages to avoid orphaned parts